+++ testcontainers-selenium/src/test/java/software/xdev/testcontainers/selenium/containers/browser/GenericContainerImageNameAccessorTest.java
@@ -1,0 +1,52 @@
+/*
+ * Copyright © 2024 XDEV Software (https://xdev.software)
+ *
+ * Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ *
+ *     http://www.apache.org/licenses/LICENSE-2.0
+ *
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
+package software.xdev.testcontainers.selenium.containers.browser;
+
+import static org.junit.jupiter.api.Assertions.assertDoesNotThrow;
+import static org.junit.jupiter.api.Assertions.assertEquals;
+
+import java.util.concurrent.CompletableFuture;
+import java.util.stream.Stream;
+
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.Arguments;
+import org.junit.jupiter.params.provider.MethodSource;
+import org.testcontainers.containers.GenericContainer;
+import org.testcontainers.utility.DockerImageName;
+
+
+class GenericContainerImageNameAccessorTest
+{
+	static final String IMAGE = "alpine:3";
+	static final DockerImageName DOCKER_IMAGE_NAME = DockerImageName.parse(IMAGE);
+	
+	@ParameterizedTest
+	@MethodSource
+	void reflectionStillWorks(final GenericContainer<?> container)
+	{
+		assertEquals(
+			DOCKER_IMAGE_NAME,
+			assertDoesNotThrow(() -> GenericContainerImageNameAccessor.getImageName(container)));
+	}
+	
+	static Stream<Arguments> reflectionStillWorks()
+	{
+		return Stream.of(
+				new GenericContainer<>(IMAGE),
+				new GenericContainer<>(CompletableFuture.completedFuture(IMAGE)))
+			.map(Arguments::of);
+	}
+}