+++ CHANGELOG.md
@@ -1,3 +1,9 @@
+# 2.1.0
+* Rename `SeleniumUtils` to `SeleniumVersionDetector`
+* Extract image name extraction logic
+* Update default Selenium version to `4.47.0`
+* Updated dependencies
+
 # 2.0.2
 * Extract wait strategy values into constants
 * Updated dependencies